Move and expose types in the enzyme-matchers package
