+++ packages/enzyme-matchers/package.json
@@ -18,8 +18,7 @@
     "prepublish": "npm run build"
   },
   "files": [
-    "lib",
-    "types"
+    "lib"
   ],
   "repository": {
     "type": "git",
