+++ packages/enzyme-matchers/src/assertions/toHaveText.js
@@ -6,9 +6,7 @@
  * @flow
  */
 
-import type { Matcher } from '../../../../types/Matcher';
-import type { EnzymeObject } from '../../../../types/EnzymeObject';
-
+import type { EnzymeObject, Matcher } from '../types';
 import name from '../utils/name';
 import html from '../utils/html';
 import single from '../utils/single';
