+++ packages/enzyme-matchers/src/types/MatcherMethods.js
