+++ packages/enzyme-matchers/src/types/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export type { EnzymeObject } from './EnzymeObject';
+export type { Matcher } from './Matcher';
+export type { MatcherMethods } from './MatcherMethods';
